+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 * fix: UserScript `registerMenuCommand` does not remove old listeners when re-registering a menu with an existing `id`
 * fix: Notification support (misused `id`)
+* fix: Keep the background page alive in Firefox (offscreen API not yet supported)
 
 # 3.7.3
 * fix: UserScript `run-at` "panel" value not supported with Firefox (running them in `document_idle` instead)